refactor: split package into modules
