+++ cmd/talia/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"os"
+
+	"github.com/sustanza/talia"
+)
+
+func main() {
+	os.Exit(talia.RunCLI(os.Args[1:]))
+}
